refactor(videohub): abstract out duplicate code
