+++ constants/index.js
@@ -94,6 +94,27 @@ export const SUB_BRAND_LINKS = [
   READR_LINK,
 ]
 
+export const VIDEOHUB_CATEGORIES_PLAYLIST_MAPPING = {
+  // 鏡封面
+  video_coverstory: 'PLftq_bkhPR3ZtDGBhyqVGObQXazG_O3M3',
+  // 鏡娛樂
+  video_entertainment: 'PLftq_bkhPR3aj8UaqBvel6wia54AM5wlh',
+  // 鏡社會
+  video_society: 'PLftq_bkhPR3bLVBh5khl2pLxgFoPwrbfl',
+  // 鏡調查
+  video_investigation: 'PLftq_bkhPR3YOrSnIpcqSkY3hPE2TjXfW',
+  // 鏡人物
+  video_people: 'PLftq_bkhPR3YkNjH8VQZ__8nXZ9INIjAu',
+  // 鏡財經
+  video_finance: 'PLftq_bkhPR3afBv0Wg_oUqjd_pkWIJm2h',
+  // 鏡食旅
+  video_foodtravel: 'PLftq_bkhPR3baCfd6RU_1hbkY8ynXssun',
+  // 娛樂透視
+  video_ent_perspective: 'PLftq_bkhPR3YxUNEIHIMA2fsM-DqxCHMb',
+  // 汽車鐘錶
+  video_carandwatch: 'PLgvIJQ8OtT8LOdwVF4P9hdQiuf6uAiwb6',
+}
+
 export const SITE_TITLE = '鏡週刊 Mirror Media'
 export const SITE_URL = 'https://www.mirrormedia.mg'
 export const SITE_OG_IMAGE =
