+++ pages/__tests__/section.videohub.spec.js
@@ -49,70 +49,6 @@ describe('isThisSection method', () => {
   })
 })
 
-describe('isValidYoutubeVideo method', () => {
-  const wrapper = createWrapper(videohub)
-  test('video item with normal title is valid', () => {
-    expect(wrapper.vm.isValidYoutubeVideo({ title: 'test' })).toBe(true)
-  })
-  test('video item with "Deleted video" title is invalid', () => {
-    expect(wrapper.vm.isValidYoutubeVideo({ title: 'Deleted video' })).toBe(
-      false
-    )
-  })
-  test('video item with "Private video" title is invalid', () => {
-    expect(wrapper.vm.isValidYoutubeVideo({ title: 'Private video' })).toBe(
-      false
-    )
-  })
-})
-
-describe('restructureItem method', () => {
-  const wrapper = createWrapper(videohub)
-  const videoIdMock = 'POYXvnGlPt8'
-  const titleMock = 'test title'
-  const thumbnailsUrlMock = 'test url'
-  test('should return the proper restuctured object', () => {
-    const origin = {
-      id: {
-        videoId: videoIdMock,
-      },
-      snippet: {
-        title: titleMock,
-        thumbnails: {
-          high: {
-            url: thumbnailsUrlMock,
-          },
-        },
-      },
-    }
-    expect(wrapper.vm.restructureItem(origin)).toEqual({
-      videoId: videoIdMock,
-      title: titleMock,
-      thumbnails: thumbnailsUrlMock,
-    })
-  })
-  test('should return the proper restuctured object', () => {
-    const origin = {
-      snippet: {
-        title: titleMock,
-        resourceId: {
-          videoId: videoIdMock,
-        },
-        thumbnails: {
-          high: {
-            url: thumbnailsUrlMock,
-          },
-        },
-      },
-    }
-    expect(wrapper.vm.restructureItem(origin)).toEqual({
-      videoId: videoIdMock,
-      title: titleMock,
-      thumbnails: thumbnailsUrlMock,
-    })
-  })
-})
-
 describe('handleSendGA method', () => {
   const $ga = {
     event: jest.fn(),
