+++ pages/section/videohub.vue
@@ -50,7 +50,12 @@
 <script>
 import { mapState } from 'vuex'
 import _ from 'lodash'
-import { SITE_TITLE, SITE_URL } from '~/constants/index'
+import {
+  SITE_TITLE,
+  SITE_URL,
+  VIDEOHUB_CATEGORIES_PLAYLIST_MAPPING as PLAYLIST_MAPPING,
+} from '~/constants/index'
+import { processResponseItems as processItems } from '~/utils/youtube'
 import ContainerFullScreenAds from '~/components/ContainerFullScreenAds.vue'
 import UIStickyAd from '~/components/UIStickyAd.vue'
 import UIVideoCategory from '~/components/UIVideoCategory.vue'
@@ -58,27 +63,6 @@ import UIVideoIframeWithItems from '~/components/UIVideoIframeWithItems.vue'
 import UIVideoPopular from '~/components/UIVideoPopular.vue'
 import UIVideoSubscriptions from '~/components/UIVideoSubscriptions.vue'
 import gptUnits from '~/constants/gptUnits'
-
-const PLAYLIST_MAPPING = {
-  // 鏡封面
-  video_coverstory: 'PLftq_bkhPR3ZtDGBhyqVGObQXazG_O3M3',
-  // 鏡娛樂
-  video_entertainment: 'PLftq_bkhPR3aj8UaqBvel6wia54AM5wlh',
-  // 鏡社會
-  video_society: 'PLftq_bkhPR3bLVBh5khl2pLxgFoPwrbfl',
-  // 鏡調查
-  video_investigation: 'PLftq_bkhPR3YOrSnIpcqSkY3hPE2TjXfW',
-  // 鏡人物
-  video_people: 'PLftq_bkhPR3YkNjH8VQZ__8nXZ9INIjAu',
-  // 鏡財經
-  video_finance: 'PLftq_bkhPR3afBv0Wg_oUqjd_pkWIJm2h',
-  // 鏡食旅
-  video_foodtravel: 'PLftq_bkhPR3baCfd6RU_1hbkY8ynXssun',
-  // 娛樂透視
-  video_ent_perspective: 'PLftq_bkhPR3YxUNEIHIMA2fsM-DqxCHMb',
-  // 汽車鐘錶
-  video_carandwatch: 'PLgvIJQ8OtT8LOdwVF4P9hdQiuf6uAiwb6',
-}
 
 const INVERTED_PLAYLIST_MAPPING = _.invert(PLAYLIST_MAPPING)
 
@@ -94,7 +78,7 @@ export default {
   },
   async fetch() {
     const response = await this.fetchChannelData({ order: 'date' })
-    this.latestData = this.processItems(response).slice(0, 5)
+    this.latestData = processItems(response).slice(0, 5)
   },
   data() {
     return {
@@ -138,7 +122,7 @@ export default {
     },
     async fetchPopularData() {
       const response = await this.fetchChannelData({ order: 'viewCount' })
-      this.popularData = this.processItems(response)
+      this.popularData = processItems(response)
     },
     fetchChannelData({ order = 'date' } = {}) {
       return this.$fetchYoutubeSearch({
@@ -166,11 +150,6 @@ export default {
       const currentSectionName = this.$route.path.split('/section/')[1]
       return section.name === currentSectionName
     },
-    isValidYoutubeVideo(item) {
-      // for specific title from Youtube response data
-      const invalidTitles = ['Deleted video', 'Private video']
-      return !invalidTitles.includes(item.title)
-    },
     mapCategoryToPlatlistId(category) {
       return PLAYLIST_MAPPING[category.name]
     },
@@ -180,19 +159,8 @@ export default {
         this.$set(
           this.categoriesPlaylistData,
           categoryName,
-          this.processItems(data.value)
+          processItems(data.value)
         )
-      }
-    },
-    processItems(response = {}) {
-      const items = response.items ?? []
-      return items.map(this.restructureItem).filter(this.isValidYoutubeVideo)
-    },
-    restructureItem(item) {
-      return {
-        videoId: item.id?.videoId || item.snippet?.resourceId?.videoId,
-        title: item.snippet?.title,
-        thumbnails: item.snippet?.thumbnails?.high?.url,
       }
     },
   },
